connect content moved into index
